Tweaked which mailer variables belong in `.env`
After doing some 🤔, realized that `MAILER_FROM_ADDRESS` is not
environment-specific, and also that we should keep our `.env.example`
entirely clean of real configuration. I also organized some
component-specific constants out of the global constants file.

* Moved `MAILER_FROM_ADDRESS` to `workers/mailer/constants.js`
* Moved the two `MAILING_LIST` constants to `newsletters/constants.js`
* Removed real domain configuration out of `.env.example`

Affects #23
