+++ .env.example
@@ -13,13 +13,10 @@ REDIS_URL=redis://127.0.0.1:6379
 # A value of `*` will load all queues.
 ENABLED_QUEUE_DIRECTORIES=*
 
-# Generic mailer settings
-MAILER_FROM_ADDRESS='Tech & Check Alerts <alerts@factstream.co>'
-
 # Mailgun-specific settings
 # (Get these from Mailgun)
-MAILGUN_API_DOMAIN=alerts.factstream.co
-MAILGUN_API_URL=https://api.mailgun.net/v3/alerts.factstream.co
+MAILGUN_API_DOMAIN=domain.org
+MAILGUN_API_URL=https://api.mailgun.net/v3/domain.org
 MAILGUN_API_KEY=ABC123
 
 # The amount of detail you want to appear in logs
