+++ src/server/constants.js
@@ -4,18 +4,4 @@ export const ENV_NAMES = {
   PRODUCTION: 'production',
 }
 
-export const MAILING_LISTS = {
-  DEVELOPERS: 'DEVELOPERS',
-  TESTERS: 'TESTERS',
-  PRIMARY: 'PRIMARY',
-  NORTH_CAROLINA: 'NORTH_CAROLINA',
-}
-
-export const MAILING_LIST_ADDRESSES = {
-  DEVELOPERS: 'dev@alerts.factstream.co',
-  TESTERS: 'test@alerts.factstream.co',
-  PRIMARY: 'alerts@alerts.factstream.co',
-  NORTH_CAROLINA: 'northcarolina@alerts.factstream.co',
-}
-
 export const CLAIMBUSTER_THRESHHOLD = 0.5
