+++ src/server/newsletters/constants.js
@@ -1,0 +1,13 @@
+export const MAILING_LISTS = {
+  DEVELOPERS: 'DEVELOPERS',
+  TESTERS: 'TESTERS',
+  PRIMARY: 'PRIMARY',
+  NORTH_CAROLINA: 'NORTH_CAROLINA',
+}
+
+export const MAILING_LIST_ADDRESSES = {
+  DEVELOPERS: 'dev@alerts.factstream.co',
+  TESTERS: 'test@alerts.factstream.co',
+  PRIMARY: 'alerts@alerts.factstream.co',
+  NORTH_CAROLINA: 'northcarolina@alerts.factstream.co',
+}
