+++ src/server/workers/mailer/constants.js
@@ -1,0 +1,4 @@
+// Disabling because we intend to have more exports in the future.
+/* eslint-disable import/prefer-default-export */
+
+export const MAILER_FROM_ADDRESS = 'Tech & Check Alerts <alerts@factstream.co>'
